fix: deleting cloud files only if the path exists (#44)
* fix: delete cloud files only if the path exists

- `storage.listdir(path)` throws an error for the path As for the first time path doesn't exists on S3. Error: `NoSuchKey: An error occurred (NoSuchKey) when calling the ListObjects operation: The specified key does not exist`.
- In this commit, added a check if the specific path does exists or not before deleting the previously added content on that path.

* fix: CI failure occurring due to Ubuntu 24 runner.

- CI is failing here: https://github.com/edly-io/h5pxblock/actions/runs/13648230543/job/38151064603 and it's explanation can be found here: actions/runner-images#11666 (comment) along with it's fix.

* chore: version bump

---------

+++ h5pxblock/utils.py
@@ -63,17 +63,18 @@ def delete_existing_files_cloud(storage, path):
     """
     Recusively delete all files under given path on cloud storage
     """
-    log.info("%s path is being deleted on cloud", path)
-    dir_names, file_names = storage.listdir(path)
-    with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_WORKERS) as executor:
-        for file_name in file_names:
-            file_path = os.path.join(path, file_name)
-            future = executor.submit(storage.delete, file_path)
-            future.add_done_callback(future_result_handler)
+    if storage.exists(path):
+        log.info("%s path is being deleted on cloud", path)
+        dir_names, file_names = storage.listdir(path)
+        with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_WORKERS) as executor:
+            for file_name in file_names:
+                file_path = os.path.join(path, file_name)
+                future = executor.submit(storage.delete, file_path)
+                future.add_done_callback(future_result_handler)
 
-    for dir_name in dir_names:
-        dir_path = os.path.join(path, dir_name)
-        delete_existing_files_cloud(storage, dir_path)
+        for dir_name in dir_names:
+            dir_path = os.path.join(path, dir_name)
+            delete_existing_files_cloud(storage, dir_path)
 
 
 def unpack_package_local_path(package, path):
